cleanup(epic-28-r2): consolidate ~28 duplicate test doubles into shared TestDoubles namespace (PF-C4)
Round-2 fix waves had eight different agents inventing their own per-file
private sealed test doubles for the same handful of interfaces. The merge-
fixup commit (b8260c4) reconciled signatures but did not dedupe. PF-C4
flagged ~13 duplicates across `ITenantConnectionResolver`,
`IPlatformEventPublisher`, `ITenantStatusInvalidationBus`, and
`ITenantStatusCache`; full inventory across the round-2 surface turned up
~28 once `IPlatformEventRepository` and `ISecretAccessAuditor` were folded
in.

This is a pure consolidation pass — no production code touched, no test
added or removed. Behaviour is identical: every shared double exposes the
same surface (and where call-site naming diverged, *both* property names
are aliased so existing assertions compile unchanged).

## What was consolidated

8 shared doubles now live in
`apps/tamma-elsa/tests/Tamma.Api.Tests/TestDoubles/` as
`internal sealed class`:

| Shared double                          | Replaces (count) |
|----------------------------------------|-----------------:|
| RecordingTenantConnectionResolver      | 3 |
| NoopTenantConnectionResolver           | 4 |
| RecordingPlatformEventPublisher        | 7 |
| RecordingStatusCache                   | 4 |
| RecordingInvalidationBus               | 2 |
| RecordingSecretAccessAuditor           | 3 |
| RecordingPlatformEventRepository       | 3 |
| NoopPlatformEventRepository            | 2 |

Total: **28 duplicate `private sealed class` blocks removed**, replaced
with 8 shared `internal sealed class` files.

## Files updated (15)

- Admin/AdminImpersonationTests.cs
- Admin/AdminTenantsAuditAndNoteTests.cs
- Admin/AdminTenantsTests.cs
- Alerts/AlertEventEmitterTests.cs
- Auth/AuthAuditEventTests.cs
- Auth/SwitchOrgEndpointTests.cs
- Epic28/QuickWinsRound2Tests.cs
- Middleware/TenantContextMiddlewareTests.cs
- Secrets/KekRotationAdvisoryLockTests.cs
- Secrets/KekRotationCoordinatorTests.cs
- Secrets/KekRotationPostFixTests.cs
- Secrets/KekRotationRetryTests.cs
- Secrets/Query/SecretQueryServiceTests.cs
- Secrets/Reveal/SecretRevealServiceTests.cs
- Secrets/Stopgap/StopgapSecretMigratorTests.cs

Net diff: 15 files changed, 56 insertions(+), 531 deletions(-).

## Behavioural decisions

- **`RecordingTenantConnectionResolver`** is the strictly-more-defensive
  flavour: thread-safe `_evicted` list with snapshot accessor, throws
  `NotImplementedException` on every non-evict method. Exposes both
  `Evictions` (mutable list — Admin/KEK call sites) and `Evicted`
  (snapshot list — concurrency-sensitive call sites) so neither convention
  has to change. `GetStats()` returns `new(0,0,0)` rather than `default`
  (which is `null` for a record class) — strictly safer; AdminTenantsTests
  never inspected the value, so flipping null→non-null is invisible.
- **`NoopTenantConnectionResolver`** keeps EvictAsync silent (no
  recording) for tests that don't care about pool eviction tracking.
  Distinct class so the test's intent stays visible at the construction
  site.
- **`RecordingPlatformEventPublisher`** auto-assigns Id+CreatedAt on
  append (matching the real Postgres repository) — strictly more
  conservative than the QuickWins variant which was passive. Exposes
  both `Events` and `Appended` aliases for the same backing list.
- **`RecordingStatusCache`** is the most-functional flavour
  (`Entries` dictionary + `Reads` + `Invalidations`). The slim
  `NoopStatusCache` and `Set/Invalidate-only` variants are subsumed
  because tests that don't pre-seed `Entries` see identical
  `TryGet => false` semantics.

## Doubles intentionally left local

- **`TenantStatusInvalidationListenerTests.RecordingResolver`** +
  **`SlowResolver`** — concurrency-sensitive Postgres LISTEN/NOTIFY
  integration test with thread-coordinating `TaskCompletionSource`. The
  shape diverges enough that consolidation would obscure intent.
- **`KekRotationRetryTests.RecordingRedactor`** — single-use double for
  `IErrorRedactor`; no other test exercises the same shape.
- **`InMemoryPlatformEventBusTests.StubDedupRepository`** — purpose-built
  to return null on append (simulating the partial-unique-index dedup
  hit). Distinct semantic from the recording/noop pattern.
- **5 `RecordingEmitter`/`RecordingAlertEmitter` copies** in
  `Tamma.Activities.Tests` — each exposes a different `Events` typed
  list (Budget vs AgentDispatch vs Workflow vs Unhealthy vs
  SecretRotation). Consolidation would require renaming `Events` →
  typed property at every call site, violating "keep call sites
  identical". Net consolidation win there is small enough to defer.
- **`AlertEventEmitterTests.RecordingEventRepository`** — implements
  `IEventRepository` (tenant DCB store), distinct from
  `IPlatformEventRepository`. Single-use; not consolidated.

## Test results

3267 / 0 / 3 — identical to the integration/epic-28-r2-fixes baseline:
- Tamma.Core.Tests:        23 / 0 / 0
- Tamma.Activities.Tests: 1129 / 0 / 0
- Tamma.Api.Tests:        2115 / 0 / 3 (3 long-running skipped)

Closes the last hygiene item from the Epic 28 round-2 post-fix review
(docs/review/epic-28-round2-postfix-review-2026-04-26.md).
